Add planned change
Planning on adding an option to output all of the commands as it does now via "?" and/or "help" and/or "h". 
Instead of just displaying every option every time the welcome function is run, it'll just output when you need
to remember one of the commands. This probably won't take long to implement. If anyone would rather it output
every time then let me know and I won't make this change.